+++ sources/leddevice/dev_spi/LedDeviceAPA104.cpp
@@ -55,14 +55,14 @@ LedDevice* LedDeviceAPA104::construct(const QJsonObject& deviceConfig)
 
 bool LedDeviceAPA104::init(const QJsonObject& deviceConfig)
 {
-	_baudRate_Hz = 2235000;
+//	_baudRate_Hz = 2235000;
 
 	bool isInitOK = false;
 
 	// Initialise sub-class
 	if (ProviderSpi::init(deviceConfig))
 	{
-		WarningIf((_baudRate_Hz < 2000000 || _baudRate_Hz > 2470000), _log, "SPI rate %d outside recommended range (2000000 -> 2470000)", _baudRate_Hz);
+//		WarningIf((_baudRate_Hz < 2000000 || _baudRate_Hz > 2470000), _log, "SPI rate %d outside recommended range (2000000 -> 2470000)", _baudRate_Hz);
 
 		_ledBuffer.resize(_ledRGBCount * SPI_BYTES_PER_COLOUR + SPI_FRAME_END_LATCH_BYTES, 0x00);
