--- sources/leddevice/dev_spi/LedDeviceAWA_spi.cpp
@@ -1,196 +1,0 @@
-#include "LedDeviceAWA_spi.h"
-#include <sys/ioctl.h>
-#include <QtEndian>
-
-LedDeviceAWA_spi::LedDeviceAWA_spi(const QJsonObject& deviceConfig)
-	: ProviderSpi(deviceConfig)
-{
-	_headerSize = 6;
-
-	_white_channel_calibration = false;
-	_white_channel_limit = 255;
-	_white_channel_red = 255;
-	_white_channel_green = 255;
-	_white_channel_blue = 255;
-    _spiType = "";
-}
-
-LedDevice* LedDeviceAWA_spi::construct(const QJsonObject& deviceConfig)
-{
-	return new LedDeviceAWA_spi(deviceConfig);
-}
-
-bool LedDeviceAWA_spi::init(const QJsonObject& deviceConfig)
-{
-	bool isInitOK = false;
-
-	// Initialise sub-class
-	if (ProviderSpi::init(deviceConfig))
-	{
-		_white_channel_calibration = deviceConfig["white_channel_calibration"].toBool(false);
-		_white_channel_limit = qMin(qRound(deviceConfig["white_channel_limit"].toDouble(1) * 255.0 / 100.0), 255);
-		_white_channel_red = qMin(deviceConfig["white_channel_red"].toInt(255), 255);
-		_white_channel_green = qMin(deviceConfig["white_channel_green"].toInt(255), 255);
-		_white_channel_blue = qMin(deviceConfig["white_channel_blue"].toInt(255), 255);
-        _spiType = deviceConfig["spitype"].toString(_spiType);
-
-        Debug(_log, "_spiType: %s", QSTRING_CSTR(_spiType));
-
-		CreateHeader();
-
-		if (_white_channel_calibration)
-			Debug(_log, "White channel limit: %i, red: %i, green: %i, blue: %i", _white_channel_limit, _white_channel_red, _white_channel_green, _white_channel_blue);
-
-		isInitOK = true;
-	}
-
-	return isInitOK;
-}
-
-void LedDeviceAWA_spi::CreateHeader()
-{
-	unsigned int totalLedCount = _ledCount - 1;
-
-	auto finalSize = (uint64_t)_headerSize + (_ledCount * 3) + 8;
-	_ledBuffer.resize(finalSize, 0x00);
-
-	Debug(_log, "SPI driver with data integration check AWA protocol");
-
-	_ledBuffer[0] = 'A';
-	_ledBuffer[1] = 'w';
-	_ledBuffer[2] = (_white_channel_calibration) ? 'A' : 'a';
-	qToBigEndian<quint16>(static_cast<quint16>(totalLedCount), &_ledBuffer[3]);
-	_ledBuffer[5] = _ledBuffer[3] ^ _ledBuffer[4] ^ 0x55; // Checksum
-
-	Debug(_log, "SPI header for %d leds: %c%c%c 0x%02x 0x%02x 0x%02x", _ledCount,
-		_ledBuffer[0], _ledBuffer[1], _ledBuffer[2], _ledBuffer[3], _ledBuffer[4], _ledBuffer[5]);
-}
-
-int LedDeviceAWA_spi::write(const std::vector<ColorRgb>& ledValues)
-{
-	if (_ledCount != ledValues.size())
-	{
-		Warning(_log, "AWA spi led's number has changed (old: %d, new: %d). Rebuilding buffer.", _ledCount, ledValues.size());
-
-		_ledCount = ledValues.size();
-
-		CreateHeader();
-	}
-
-	auto bufferLength = _headerSize + ledValues.size() * sizeof(ColorRgb) + 8;
-
-	if (bufferLength > _ledBuffer.size())
-	{
-		Warning(_log, "AWA SPI buffer's size has changed. Skipping refresh.");
-
-		return 0;
-	}
-
-	uint8_t* writer = _ledBuffer.data() + _headerSize;
-	uint8_t* hasher = writer;
-
-	memcpy(writer, ledValues.data(), ledValues.size() * sizeof(ColorRgb));
-	writer += ledValues.size() * sizeof(ColorRgb);
-
-	whiteChannelExtension(writer);
-
-	uint16_t fletcher1 = 0, fletcher2 = 0, fletcherExt = 0;
-	uint8_t position = 0;
-	while (hasher < writer)
-	{
-		fletcherExt = (fletcherExt + (*(hasher) ^ (position++))) % 255;
-		fletcher1 = (fletcher1 + *(hasher++)) % 255;
-		fletcher2 = (fletcher2 + fletcher1) % 255;
-	}
-	*(writer++) = (uint8_t)fletcher1;
-	*(writer++) = (uint8_t)fletcher2;
-	*(writer++) = (uint8_t)((fletcherExt != 0x41) ? fletcherExt : 0xaa);
-	bufferLength = writer -_ledBuffer.data();
-
-	if (_spiType == "esp8266")
-		return writeBytesEsp8266(bufferLength, _ledBuffer.data());
-	else if (_spiType == "esp32")
-		return writeBytesEsp32(bufferLength, _ledBuffer.data());
-	else
-		return writeBytes(bufferLength, _ledBuffer.data());
-}
-
-void LedDeviceAWA_spi::whiteChannelExtension(uint8_t*& writer)
-{
-	if (_white_channel_calibration)
-	{
-		*(writer++) = _white_channel_limit;
-		*(writer++) = _white_channel_red;
-		*(writer++) = _white_channel_green;
-		*(writer++) = _white_channel_blue;
-	}
-}
-
-
-int LedDeviceAWA_spi::writeBytesEsp8266(unsigned size, const uint8_t* data)
-{
-    uint8_t* startData = (uint8_t*)data;
-    uint8_t* endData = (uint8_t*)data + size;
-    uint8_t buffer[34];
-
-    if (_fid < 0)
-    {
-        return -1;
-    }
-
-    _spi.tx_buf = __u64(&buffer);
-    _spi.len = __u32(34);
-    _spi.delay_usecs = 0;
-
-    int retVal = 0;
-
-    while (retVal >= 0 && startData < endData)
-    {
-        memset(buffer, 0, sizeof(buffer));
-        buffer[0] = 2;
-        buffer[1] = 0;
-        for (int i = 0; i < 32 && startData < endData; i++, startData++)
-        {
-            buffer[2 + i] = *startData;
-        }
-        retVal = ioctl(_fid, SPI_IOC_MESSAGE(1), &_spi);
-        ErrorIf((retVal < 0), _log, "SPI failed to write. errno: %d, %s", errno, strerror(errno));
-    }
-
-    return retVal;
-}
-
-int LedDeviceAWA_spi::writeBytesEsp32(unsigned size, const uint8_t* data)
-{
-    static const int      REAL_BUFFER = 1536;
-    static const uint32_t BUFFER_SIZE = REAL_BUFFER + 8;
-
-    uint8_t* startData = (uint8_t*)data;
-    uint8_t* endData = (uint8_t*)data + size;
-    uint8_t buffer[BUFFER_SIZE];
-
-    if (_fid < 0)
-    {
-        return -1;
-    }
-
-    _spi.tx_buf = __u64(&buffer);
-    _spi.len = __u32(BUFFER_SIZE);
-    _spi.delay_usecs = 0;
-
-    int retVal = 0;
-
-    while (retVal >= 0 && startData < endData)
-    {
-        memset(buffer, 0, sizeof(buffer));
-        for (int i = 0; i < REAL_BUFFER && startData < endData; i++, startData++)
-        {
-            buffer[i] = *startData;
-        }
-        buffer[REAL_BUFFER] = 0xAA;
-        retVal = ioctl(_fid, SPI_IOC_MESSAGE(1), &_spi);
-        ErrorIf((retVal < 0), _log, "SPI failed to write. errno: %d, %s", errno, strerror(errno));
-    }
-
-    return retVal;
-}
