+++ sources/leddevice/dev_spi/LedDeviceWs2812SPI.cpp
@@ -54,14 +54,14 @@ LedDevice* LedDeviceWs2812SPI::construct(const QJsonObject& deviceConfig)
 
 bool LedDeviceWs2812SPI::init(const QJsonObject& deviceConfig)
 {
-	_baudRate_Hz = 2600000;
+//	_baudRate_Hz = 2600000;
 
 	bool isInitOK = false;
 
 	// Initialise sub-class
 	if (ProviderSpi::init(deviceConfig))
 	{
-		WarningIf((_baudRate_Hz < 2106000 || _baudRate_Hz > 3075000), _log, "SPI rate %d outside recommended range (2106000 -> 3075000)", _baudRate_Hz);
+//		WarningIf((_baudRate_Hz < 2106000 || _baudRate_Hz > 3075000), _log, "SPI rate %d outside recommended range (2106000 -> 3075000)", _baudRate_Hz);
 
 		_ledBuffer.resize(_ledRGBCount * SPI_BYTES_PER_COLOUR + SPI_FRAME_END_LATCH_BYTES, 0x00);
